removed redundant options
The only way for a user to be identified as an admin, or creator, is to be logged in in the first place, so I find it confusing to speak about 3 possibilities, where actually the two of them are contained in the first one. I scrolled down the tutorial to see what logic will actually be implemented and it seems that indeed, the only checks performed are for the user to be logged in.